Rewrite _viewer dispatchers upon MacOS install
Even if the libtbx_env environment is pointing to the correct executable,
this is rewritten on cold reconfiguration (which the installer does once
in-place on the target system).

This replaces the called executable once in-place, and after reconfiguration,
upon system installation. This will not preserve this path on reconfiguration,
which isn't normally an action we support for install bundles.

Works around cctbx/cctbx_project#739.